Moved CollapsibleSection.tsx to /commons
