+++ src/components/index/BlocksList.tsx
@@ -5,7 +5,7 @@ import { Link } from '@components/commons/link/Link'
 import { TextMiddleTruncate } from '@components/commons/TextMiddleTruncate'
 import { AddressLink } from '@components/commons/link/AddressLink'
 import { MdOutlineKeyboardArrowDown, MdOutlineKeyboardArrowUp, MdStairs } from 'react-icons/md'
-import { CollapsibleSection } from '@components/vaults/common/VaultCollapsibleSection'
+import { CollapsibleSection } from '@components/commons/CollapsibleSection'
 import { Transition } from '@headlessui/react'
 import React, { useState } from 'react'
 import { BlockLink } from '@components/commons/link/BlockLink'
